More refactoring of existing code
Moved shared code specific to the generator out of common/helpers.ts and
into a generator/helpers.ts file.
Moved all non-shared code out of common/helpers.ts into respsective
source files where it was used.
