+++ src/generator/helpers.ts
@@ -1,0 +1,56 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { Session } from '@azure-tools/autorest-extension-base';
+import { comment } from '@azure-tools/codegen';
+import { CodeModel, Language, Parameter } from '@azure-tools/codemodel';
+import { ParamInfo } from '../common/helpers';
+
+
+// returns the common source-file preamble (license comment, package name etc)
+export async function contentPreamble(session: Session<CodeModel>): Promise<string> {
+  const headerText = comment(await session.getValue("header-text", "MISSING LICENSE HEADER"), "// ");
+  const namespace = await session.getValue('namespace');
+  let text = `${headerText}\n\n`;
+  text += `package ${namespace}\n\n`;
+  return text;
+}
+
+// returns true if the language contains a description
+export function hasDescription(lang: Language): boolean {
+  return (lang.description !== undefined && lang.description.length > 0 && !lang.description.startsWith('MISSING'));
+}
+
+// used to sort strings in ascending order
+export function sortAscending(a: string, b: string): number {
+  return a < b ? -1 : a > b ? 1 : 0;
+}
+
+// returns the type name with possible * prefix
+export function formatParamInfoTypeName(param: ParamInfo): string {
+  if (param.required) {
+    return param.type;
+  }
+  return `*${param.type}`;
+}
+
+// returns true if the parameter should not be URL encoded
+export function skipURLEncoding(param: Parameter): boolean {
+  if (param.extensions) {
+    return param.extensions['x-ms-skip-url-encoding'] === true;
+  }
+  return false;
+}
+
+// sorts ParamInfo objects by their required state, ordering required before optional
+export function sortParamInfoByRequired(a: ParamInfo, b: ParamInfo): number {
+  if (a.required === b.required) {
+    return 0;
+  }
+  if (a.required && !b.required) {
+    return -1;
+  }
+  return 1;
+}
