+++ src/generator/models.ts
@@ -7,7 +7,9 @@ import { Session } from '@azure-tools/autorest-extension-base';
 import { comment, pascalCase } from '@azure-tools/codegen';
 import { CodeModel, ConstantSchema, ImplementationLocation, ObjectSchema, Language, Schema, SchemaType, Parameter, Property } from '@azure-tools/codemodel';
 import { values } from '@azure-tools/linq';
-import { contentPreamble, hasDescription, ImportManager, isArraySchema, sortAscending } from '../common/helpers';
+import { isArraySchema } from '../common/helpers';
+import { contentPreamble, hasDescription, sortAscending } from './helpers';
+import { ImportManager } from './imports';
 
 // Creates the content in models.go
 export async function generateModels(session: Session<CodeModel>): Promise<string> {
